Adding basic pytest testcases
